test: #19 DeleteNoticeUseCase 단위, 통합 테스트 추가

+++ src/main/java/com/rsupport/notice/application/notice/usecase/DeleteNoticeUseCase.java
@@ -23,6 +23,10 @@ public class DeleteNoticeUseCase {
         noticeService.deleteNotice(noticeId, now);
 
         List<NoticeFile> deletedNoticeFileList = noticeFileService.getNoticeFileList(noticeId);
+        if(deletedNoticeFileList.isEmpty()) {
+            return;
+        }
+
         noticeFileService.deleteNoticeFiles(noticeId, now);
 
         for (NoticeFile noticeFile : deletedNoticeFileList) {
